varous little fixes in stop/start and commands

+++ stop.sh
@@ -1,7 +1,8 @@
 #!/bin/sh
 
 PIDFILE=kippo.pid
-PID=$(cat $PIDFILE)
-
-echo "Stopping Kippo .."
-kill -TERM $PID
+PID=$(cat $PIDFILE 2>/dev/null)
+if [ -n "$PID" ]; then
+    echo "Stopping Kippo .."
+    kill -TERM $PID
+fi
